Revise 48ac4d1 to fix DNS syncing
It looks like there was an issue where "exchange: 100 mx.sendgrid.net."
should have been "exchange: mx.sendgrid.net.". OctoDNS should have
caught this with its tests, but missed it.

Please note: if the "100" was intended to specify the preference (AKA
priority) for the MX record, the preference value should be updated from
"1" to "100" in the YAML.

I filed https://github.com/github/octodns/issues/600 in OctoDNS so they
can hopefully fix this.